+++ pages/api/matchday.js
@@ -1,4 +1,4 @@
-import { data } from '../../../../data/data.js'
+import { data } from '../../data/data.js'
 
 export default async (req, res) => {
     res.status(200).json(data[0]["matchday"]);
